+++ service/clothes_info.py
